ref(tests): Rework structure of integration tests
