+++ etl-api/tests/integration/images_test.rs
@@ -5,21 +5,8 @@ use etl_api::routes::images::{
 use etl_telemetry::tracing::init_test_tracing;
 use reqwest::StatusCode;
 
-use crate::common::test_app::{TestApp, spawn_test_app};
-
-pub async fn create_default_image(app: &TestApp) -> i64 {
-    create_image_with_name(app, "some/image".to_string(), true).await
-}
-
-pub async fn create_image_with_name(app: &TestApp, name: String, is_default: bool) -> i64 {
-    let image = CreateImageRequest { name, is_default };
-    let response = app.create_image(&image).await;
-    let response: CreateImageResponse = response
-        .json()
-        .await
-        .expect("failed to deserialize response");
-    response.id
-}
+use crate::common::mocks::create_image_with_name;
+use crate::common::test_app::spawn_test_app;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn image_can_be_created() {
