+++ etl-api/tests/integration/tenants_sources_test.rs
@@ -5,7 +5,7 @@ use etl_telemetry::tracing::init_test_tracing;
 
 use crate::{
     common::test_app::spawn_test_app,
-    integration::sources_test::{new_name, new_source_config},
+    common::mocks::sources::{new_name, new_source_config},
 };
 
 #[tokio::test(flavor = "multi_thread")]
