+++ etl-destinations/Cargo.toml
@@ -2,6 +2,7 @@
 name = "etl-destinations"
 version = "0.1.0"
 edition = "2024"
+autotests = false
 
 [features]
 bigquery = [
@@ -45,3 +46,13 @@ reqwest = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 uuid = { workspace = true, features = ["v4"] }
+
+[[test]]
+name = "pipeline_with_bigquery"
+path = "tests/pipeline_with_bigquery.rs"
+required-features = ["bigquery"]
+
+[[test]]
+name = "pipeline_with_iceberg"
+path = "tests/pipeline_with_iceberg.rs"
+required-features = ["iceberg"]
