--- etl-destinations/tests/integration/mod.rs
@@ -1,4 +1,0 @@
-#[cfg(feature = "bigquery")]
-mod bigquery_test;
-#[cfg(feature = "iceberg")]
-mod iceberg_test;
