--- etl-destinations/tests/mod.rs
@@ -1,2 +1,0 @@
-mod common;
-mod integration;
