--- etl/tests/failpoints/mod.rs
@@ -1,1 +1,0 @@
-mod pipeline_test;
