--- etl/tests/integration/mod.rs
@@ -1,4 +1,0 @@
-mod pipeline_test;
-mod postgres_store;
-mod replica_identity;
-mod replication_test;
