--- etl/tests/mod.rs
@@ -1,2 +1,0 @@
-mod failpoints;
-mod integration;
